InvalidToken: Regular Signed out screen - move class
