+++ vector/src/main/java/im/vector/riotx/features/MainActivity.kt
@@ -32,7 +32,7 @@ import im.vector.riotx.core.platform.VectorBaseActivity
 import im.vector.riotx.core.utils.deleteAllFiles
 import im.vector.riotx.features.home.HomeActivity
 import im.vector.riotx.features.login.LoginActivity
-import im.vector.riotx.features.main.SignedOutActivity
+import im.vector.riotx.features.signout.SignedOutActivity
 import kotlinx.android.parcel.Parcelize
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
